+++ example/collapse.dart
@@ -1,0 +1,8 @@
+import 'package:bootjack/bootjack.dart';
+import 'package:dquery/dquery.dart';
+
+void main() {
+  Dropdown.use();
+  Transition.use();
+  Collapse.use();
+}