+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Fixed tooltip events.
 * Allow custom HTML validator in tooltip and popover. (thanks nawafnaim)
 * Minot fixes.
+* Added collapse component
 
 **0.6.1**